fix(auth): mejorar seguridad de logout y eliminar duplicación de código
- Se añadió @require_http_methods(['POST']) en la vista de logout para prevenir vulnerabilidades CSRF.
- Se reemplazaron los enlaces de logout en las plantillas (base.html y landing_page.html) por formularios que envían solicitudes POST con CSRF tokens.
- Se consolidó el decorador no_session_required en un módulo compartido para eliminar duplicación de código en ciudadano_app y entidad_municipal_app.
- Se mantuvo la coherencia visual en los botones de cierre de sesión tras el cambio a formularios.
- Corrección de errores 405 (Method Not Allowed) al garantizar el uso de POST en todas las acciones de logout.
