+++ shared/templates/landing_page.html
@@ -25,9 +25,12 @@
               Dashboard Entidad
           </a>
       {% endif %}
-      <a href="{% url 'logout_usuario' %}" class="inline-flex items-center px-6 py-3 border border-transparent text-base font-medium rounded-md text-white bg-red-600 hover:bg-red-700">
-          Cerrar Sesión
-      </a>
+      <form action="{% url 'logout_usuario' %}" method="post" class="inline-block">
+          {% csrf_token %}
+          <button type="submit" class="inline-flex items-center px-6 py-3 border border-transparent text-base font-medium rounded-md text-white bg-red-600 hover:bg-red-700">
+              Cerrar Sesión
+          </button>
+      </form>
   {% else %}
       <a href="{% url 'login_ciudadano' %}" class="inline-flex items-center px-6 py-3 border border-transparent text-base font-medium rounded-md text-white bg-[#9D2235] hover:bg-[#B53045]">
           Acceso Ciudadano
@@ -101,12 +104,12 @@
           <div class="absolute flex items-center justify-center h-12 w-12 rounded-md bg-[#143D69] text-white">
             <!-- Icono -->
             <svg class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m5.618-4.016A11.955 11.955 0 0112 2.944a11.955 11.955 0 01-8.618 3.04A12.02 12.02 0 003 9c0 5.591 3.824 10.29 9 11.622 5.176-1.332 9-6.03 9-11.622 0-1.042-.133-2.052-.382-3.016z"/>
+              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 10V3L4 14h7v7l9-11h-7z"/>
             </svg>
           </div>
-          <p class="ml-16 text-lg leading-6 font-medium text-gray-900">Seguro y Confiable</p>
+          <p class="ml-16 text-lg leading-6 font-medium text-gray-900">Respuesta Rápida</p>
           <p class="mt-2 ml-16 text-base text-gray-500">
-            Tus datos están protegidos y seguros en todo momento.
+            Obtén respuestas rápidas a tus solicitudes y consultas.
           </p>
         </div>
       </div>
